# Conflicts:
#	.gradle/8.4/checksums/md5-checksums.bin
#	.gradle/8.4/checksums/sha1-checksums.bin
#	.gradle/8.4/dependencies-accessors/dependencies-accessors.lock
#	.gradle/8.7/fileHashes/fileHashes.bin
#	.gradle/buildOutputCleanup/buildOutputCleanup.lock
#	.gradle/file-system.probe
#	local.properties

+++ app/src/main/java/com/critetiontech/ctvitalio/utils/Utilities.java
@@ -40,7 +40,7 @@ public class Utilities {
             us_height_inc += inch * 4;
         }
 
-        gss_arg_value = (int) us_height_inc * 254;
+        gss_arg_value = us_height_inc * 254;
         gss_arg_value = (gss_arg_value + 100) / 200;
         gss_arg_value = gss_arg_value * 5;
